chore: 코드 리팩토링 및 실내 도면 카메라 이동범위 수정(#48)
