+++ src/App.jsx
@@ -7,7 +7,6 @@ import SignupPage from './pages/Signup/SignupPage'
 import MapPage from './pages/MapPage/MapPage'
 import RoutingSearchPage from './pages/RoutingSearchPage/RoutingSearchPage'
 import RoutingOptionPage from './pages/RoutingOptionPage/RoutingOptionPage'
-import TransitRouteTestPage from './pages/TransitRouteTestPage/TransitRouteTestPage'
 import SearchPage from './pages/SearchPage/SearchPage'
 import RoutingPage from './pages/RoutingPage/RoutingPage'
 import MyPage from './pages/MyPage/MyPage/MyPage'
@@ -28,7 +27,6 @@ function App() {
             <Route path={ROUTES.MAP} element={<MapPage />} />
             <Route path={ROUTES.ROUTING_SEARCH} element={<RoutingSearchPage />} />
             <Route path={ROUTES.ROUTING_OPTION} element={<RoutingOptionPage />} />
-            <Route path={ROUTES.TRANSIT_ROUTE_TEST} element={<TransitRouteTestPage />} />
             <Route path={ROUTES.ROUTING} element={<RoutingPage />} />
             <Route path={ROUTES.ROUTING_LEGACY} element={<RoutingPage />} />
             <Route path={ROUTES.SEARCH} element={<SearchPage />} />
