+++ src/constants/routes.js
@@ -5,7 +5,6 @@ export const ROUTES = {
   ROUTING_LEGACY: '/RoutingPage',
   ROUTING_OPTION: '/RoutingOptionPage',
   ROUTING_SEARCH: '/RoutingSearchPage',
-  TRANSIT_ROUTE_TEST: '/TransitRouteTestPage',
   SEARCH: '/search',
   LOGIN: '/login',
   SIGNUP: '/signup',
