--- src/mocks/map/poi.mock.js
@@ -1,100 +1,0 @@
-export const mockMapPois = [
-  {
-    id: 'dgu-new-eng-main',
-    name: '동국대 신공학관 정문',
-    address: '서울특별시 중구 필동로1길 30',
-    lat: 37.55806,
-    lng: 126.99825,
-  },
-  {
-    id: 'dgu-new-eng-1f-lobby',
-    name: '신공학관 1층 로비',
-    address: '서울특별시 중구 필동로1길 30',
-    lat: 37.55798,
-    lng: 126.99834,
-  },
-  {
-    id: 'dgu-new-eng-b1-cafe',
-    name: '신공학관 B1 카페',
-    address: '서울특별시 중구 필동로1길 30',
-    lat: 37.5579,
-    lng: 126.99818,
-  },
-  {
-    id: 'dgu-new-eng-seminar',
-    name: '신공학관 세미나실',
-    address: '서울특별시 중구 필동로1길 30',
-    lat: 37.55814,
-    lng: 126.99842,
-  },
-  {
-    id: 'dgu-library',
-    name: '중앙도서관',
-    address: '서울특별시 중구 필동로1길 30',
-    lat: 37.55831,
-    lng: 126.99786,
-  },
-  {
-    id: 'dgu-student-center',
-    name: '학생회관',
-    address: '서울특별시 중구 필동로1길 30',
-    lat: 37.55763,
-    lng: 126.99858,
-  },
-  {
-    id: 'dgu-main-gate',
-    name: '동국대학교 정문',
-    address: '서울특별시 중구 필동로1길 30',
-    lat: 37.55856,
-    lng: 126.99894,
-  },
-  {
-    id: 'dgu-hyehwa-hall',
-    name: '혜화관',
-    address: '서울특별시 중구 필동로1길 30',
-    lat: 37.55774,
-    lng: 126.99762,
-  },
-  {
-    id: 'dgu-culture-hall',
-    name: '문화관',
-    address: '서울특별시 중구 필동로1길 30',
-    lat: 37.55741,
-    lng: 126.99795,
-  },
-  {
-    id: 'dgu-vision-hall',
-    name: '법학관/만해관 인근',
-    address: '서울특별시 중구 필동로1길 30',
-    lat: 37.55867,
-    lng: 126.99735,
-  },
-  {
-    id: 'dgu-bus-stop',
-    name: '동국대입구 버스정류장',
-    address: '서울특별시 중구 장충단로 일대',
-    lat: 37.55893,
-    lng: 127.00591,
-  },
-  {
-    id: 'dgu-subway',
-    name: '동대입구역 6번 출구',
-    address: '서울특별시 중구 장충단로 201',
-    lat: 37.55819,
-    lng: 127.00549,
-  },
-  {
-    id: 'dgu-athletics',
-    name: '대운동장',
-    address: '서울특별시 중구 필동로1길 30',
-    lat: 37.55708,
-    lng: 126.99841,
-  },
-  {
-    id: 'dgu-dorm',
-    name: '기숙사(남산학사) 인근',
-    address: '서울특별시 중구 필동로1길 30',
-    lat: 37.55674,
-    lng: 126.99778,
-  },
-]
