--- src/mocks/navigation/transitRoute.mock.js
@@ -1,259 +1,0 @@
-export const mockTransitNavigationResponse = {
-  requestedDestination: {
-    x: 126.9822,
-    y: 37.5624,
-    name: '디 에스테이트',
-  },
-  routedDestination: {
-    x: 126.9822,
-    y: 37.5624,
-    name: '디 에스테이트',
-  },
-  indoor: {
-    included: true,
-    buildingName: '신세계백화점 본점',
-    floorplans: [
-      {
-        floorId: '11111111-1111-1111-1111-111111111111',
-        floorName: '1F',
-        mapImageUrl: null,
-        coordinateType: 'PIXEL',
-      },
-    ],
-  },
-  routes: [
-    {
-      routeType: 'TRANSIT',
-      routeOption: 'TRANSIT_CANDIDATE',
-      totalTimeSeconds: 2100,
-      totalDistanceMeters: 11800,
-      totalDuration: '35분',
-      failures: [],
-      legs: [
-        {
-          mode: 'WALK',
-          durationSeconds: 180,
-          distanceMeters: 220,
-          stationCount: null,
-          stops: [],
-          startName: '현재 위치',
-          endName: '서울대입구역',
-          coordinateType: 'WGS84',
-          path: [
-            { x: 126.951744, y: 37.478095, name: '현재 위치' },
-            { x: 126.95245, y: 37.47902, name: '관악로' },
-            { x: 126.95334, y: 37.48016, name: '서울대입구역' },
-          ],
-          floorSegments: [],
-          steps: [
-            {
-              instruction: '관악로를 따라 120m 직진',
-              distanceMeters: 120,
-              durationSeconds: 90,
-              x: 126.951744,
-              y: 37.478095,
-              turnType: 11,
-              mode: 'WALK',
-              streetName: '관악로',
-            },
-            {
-              instruction: '횡단보도를 건너 서울대입구역으로 이동',
-              distanceMeters: 100,
-              durationSeconds: 90,
-              x: 126.95245,
-              y: 37.47902,
-              turnType: 125,
-              mode: 'WALK',
-              streetName: '관악로',
-            },
-          ],
-        },
-        {
-          mode: 'SUBWAY',
-          routeName: '수도권 2호선',
-          transitType: 'SUBWAY',
-          durationSeconds: 1080,
-          distanceMeters: 7800,
-          stationCount: 10,
-          stops: [
-            { name: '봉천', stationId: 'subway-1', x: 126.9419, y: 37.4824 },
-            { name: '신림', stationId: 'subway-2', x: 126.9296, y: 37.4842 },
-            { name: '신대방', stationId: 'subway-3', x: 126.9131, y: 37.4875 },
-            { name: '구로디지털단지', stationId: 'subway-4', x: 126.9015, y: 37.4852 },
-            { name: '대림', stationId: 'subway-5', x: 126.8956, y: 37.4925 },
-            { name: '신도림', stationId: 'subway-6', x: 126.8912, y: 37.5087 },
-            { name: '문래', stationId: 'subway-7', x: 126.8949, y: 37.5179 },
-            { name: '영등포구청', stationId: 'subway-8', x: 126.8966, y: 37.5257 },
-          ],
-          startName: '서울대입구역',
-          endName: '을지로입구역',
-          coordinateType: 'WGS84',
-          path: [
-            { x: 126.95334, y: 37.48016, name: '서울대입구역' },
-            { x: 126.9419, y: 37.4824, name: '봉천' },
-            { x: 126.9296, y: 37.4842, name: '신림' },
-            { x: 126.9131, y: 37.4875, name: '신대방' },
-            { x: 126.9015, y: 37.4852, name: '구로디지털단지' },
-            { x: 126.8956, y: 37.4925, name: '대림' },
-            { x: 126.8912, y: 37.5087, name: '신도림' },
-            { x: 126.8949, y: 37.5179, name: '문래' },
-            { x: 126.8966, y: 37.5257, name: '영등포구청' },
-            { x: 126.9822, y: 37.5661, name: '을지로입구역' },
-          ],
-          floorSegments: [],
-          steps: [],
-        },
-        {
-          mode: 'WALK',
-          durationSeconds: 240,
-          distanceMeters: 340,
-          stationCount: null,
-          stops: [],
-          startName: '을지로입구역',
-          endName: '정문',
-          coordinateType: 'WGS84',
-          path: [
-            { x: 126.9822, y: 37.5661, name: '을지로입구역' },
-            { x: 126.9827, y: 37.5649, name: '명동길' },
-            { x: 126.9822, y: 37.5624, name: '정문' },
-          ],
-          floorSegments: [],
-          steps: [
-            {
-              instruction: '명동길을 따라 210m 직진',
-              distanceMeters: 210,
-              durationSeconds: 150,
-              x: 126.9822,
-              y: 37.5661,
-              turnType: 11,
-              mode: 'WALK',
-              streetName: '명동길',
-            },
-            {
-              instruction: '우회전 후 130m 이동',
-              distanceMeters: 130,
-              durationSeconds: 90,
-              x: 126.9827,
-              y: 37.5649,
-              turnType: 13,
-              mode: 'WALK',
-              streetName: null,
-            },
-          ],
-        },
-        {
-          mode: 'INDOOR',
-          durationSeconds: null,
-          distanceMeters: null,
-          stationCount: null,
-          stops: [],
-          startName: '정문',
-          endName: '디 에스테이트',
-          mapType: 'BUILDING',
-          mapImageUrl: null,
-          floorId: '11111111-1111-1111-1111-111111111111',
-          floorName: '1F',
-          coordinateType: 'PIXEL',
-          path: [
-            { x: 80, y: 260, name: '정문' },
-            { x: 150, y: 260, name: '중앙 통로' },
-            { x: 210, y: 220, name: '디 에스테이트' },
-          ],
-          floorSegments: [
-            {
-              mapType: 'BUILDING',
-              floorId: '11111111-1111-1111-1111-111111111111',
-              floorName: '1F',
-              mapImageUrl: null,
-              coordinateType: 'PIXEL',
-              path: [
-                { x: 80, y: 260, name: '정문' },
-                { x: 150, y: 260, name: '중앙 통로' },
-                { x: 210, y: 220, name: '디 에스테이트' },
-              ],
-              steps: [
-                {
-                  instruction: '건물 입구로 진입',
-                  distanceMeters: null,
-                  durationSeconds: null,
-                  x: 80,
-                  y: 260,
-                  turnType: null,
-                  mode: 'INDOOR',
-                  streetName: null,
-                  pathStartIndex: 0,
-                  pathEndIndex: 1,
-                },
-                {
-                  instruction: '중앙 통로를 따라 20m 직진',
-                  distanceMeters: null,
-                  durationSeconds: null,
-                  x: 150,
-                  y: 260,
-                  turnType: null,
-                  mode: 'INDOOR',
-                  streetName: null,
-                  pathStartIndex: 1,
-                  pathEndIndex: 2,
-                },
-                {
-                  instruction: '디 에스테이트 도착',
-                  distanceMeters: null,
-                  durationSeconds: null,
-                  x: 210,
-                  y: 220,
-                  turnType: null,
-                  mode: 'INDOOR',
-                  streetName: null,
-                  pathStartIndex: 1,
-                  pathEndIndex: 2,
-                },
-              ],
-            },
-          ],
-          steps: [
-            {
-              instruction: '건물 입구로 진입',
-              distanceMeters: null,
-              durationSeconds: null,
-              x: 80,
-              y: 260,
-              turnType: null,
-              mode: 'INDOOR',
-              streetName: null,
-              pathStartIndex: 0,
-              pathEndIndex: 1,
-            },
-            {
-              instruction: '중앙 통로를 따라 20m 직진',
-              distanceMeters: null,
-              durationSeconds: null,
-              x: 150,
-              y: 260,
-              turnType: null,
-              mode: 'INDOOR',
-              streetName: null,
-              pathStartIndex: 1,
-              pathEndIndex: 2,
-            },
-            {
-              instruction: '디 에스테이트 도착',
-              distanceMeters: null,
-              durationSeconds: null,
-              x: 210,
-              y: 220,
-              turnType: null,
-              mode: 'INDOOR',
-              streetName: null,
-              pathStartIndex: 1,
-              pathEndIndex: 2,
-            },
-          ],
-        },
-      ],
-    },
-  ],
-  notFoundRouteTypes: [],
-  failures: [],
-  message: null,
-}
