--- src/pages/TransitRouteTestPage/TransitRouteTestPage.css
@@ -1,46 +1,0 @@
-.transit-route-test-page {
-  min-height: 100dvh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: var(--space-24);
-  background: var(--surface-50);
-  font-family: var(--font-sans);
-}
-
-.transit-route-test-page__panel {
-  width: min(100%, 22rem);
-  display: grid;
-  gap: var(--space-14);
-  padding: var(--space-24);
-  border: var(--size-1) solid var(--gray-200);
-  border-radius: var(--radius-8);
-  background: var(--surface-0);
-  box-shadow: var(--shadow-searchbar);
-}
-
-.transit-route-test-page__panel h1 {
-  margin: 0;
-  color: var(--black-1000);
-  font-size: var(--text-20);
-  line-height: var(--line-28);
-}
-
-.transit-route-test-page__panel p {
-  margin: 0;
-  color: var(--gray-600);
-  font-size: var(--text-14);
-  line-height: var(--line-20);
-}
-
-.transit-route-test-page__panel button {
-  min-height: var(--size-48);
-  border: none;
-  border-radius: var(--radius-8);
-  background: var(--blue-600);
-  color: var(--surface-0);
-  font-family: var(--font-sans);
-  font-size: var(--text-15);
-  font-weight: 700;
-  cursor: pointer;
-}
